// types/sections.ts
export type SEODataType = {
  title: string;
  author?: string;
  description?: string;
  image?: string;
  url?: string;
  keywords: string[];
};

export type SocialLinksType = {
  [link: string]: string;
};

export type GreetingsType = {
  name: string;
  title: string;
  description?: string;
  resumeLink?: string;
};

type SoftwareSkillType = {
  skillName: string;
  iconifyTag: string;
};

type SkillType = {
  title: string;
  lottieAnimationFile: string;
  skills: React.ReactNode[] | string[];
  softwareSkills: SoftwareSkillType[];
};

export type SkillsSectionType = {
  title: string;
  subTitle: string;
  data: SkillType[];
};

export type SkillBarsType = {
  Stack: string;
  progressPercentage: string;
};

export type EducationType = {
  schoolName: string;
  subHeader: string;
  duration: string;
  desc?: string;
  grade?: string;
  descBullets?: string[];
};

export type ExperienceType = {
  role: string;
  company: string;
  companyLogo?: string;
  date: string;
  desc: string;
  descBullets?: string[];
};

export type ProjectType = {
  name: string;
  desc: string;
  github?: string;
  link?: string;
};